Document strange behavior when running as user
When I run this container using the SlimerJS driver, and specify a user that does not exist inside of the system, SlimerJS (actually XVFB) just hangs during start.  It took me a while to figure out that it's coming from the lack of a home directory, so XAuthority doesn't know where to write the files.  It's kind of an edge case, so maybe not worth changing the actual code for, but I think it should probably be documented.